[ feat ] 네이밍 import 변경 반영
