+++ src/components/signup/Role.tsx
@@ -4,8 +4,7 @@ import { styled } from "styled-components";
 import RoleCheckSignupIc from "../../assets/icon/RoleCheckSignupIc.svg";
 import RoleNoneCheckSignupIc from "../../assets/icon/RoleNoneCheckSignupIc.svg";
 import { newUserData, stepNum } from "../../atom/signup/signup";
-import { BUTTON_TEXT } from "../../core/signup/buttonText";
-import { ROLE_NAME, ROLE_SUB_TEXT, SIGNUP_TITLE } from "../../core/signup/signupTitle";
+import { BUTTON_TEXT, ROLE_NAME, ROLE_SUB_TEXT, SIGNUP_TITLE } from "../../core/signup/signUpTextLabels";
 import BackButton from "../common/BackButton";
 import BottomButton from "../common/BottomButton";
 import ProgressBar from "../common/ProgressBar";
@@ -80,7 +79,7 @@ export default function Role() {
         type="button"
         disabled={!isActive}
         isActive={isActive}
-        children={BUTTON_TEXT.done}
+        children={BUTTON_TEXT.next}
         onClick={handleDoneClick}
       />
     </>
