--- src/core/signup/buttonText.ts
@@ -1,5 +1,0 @@
-export const BUTTON_TEXT = {
-  next: "다음",
-  done: "완료",
-  signupDone: "회원가입 완료",
-};
