--- src/core/signup/signupErrorMessage.ts
@@ -1,7 +1,0 @@
-export const SIGNUP_ERROR_MESSAGE = {
-  nameError: "이름은 최소 2자 이상 입력해주세요.",
-  emailError: "올바른 이메일 형식으로 입력해 주세요.",
-  passwordError: "8 ~ 16자의 영문, 숫자, 특수문자를 모두 포함해주세요.",
-  confirmError: "비밀번호가 일치하지 않아요.",
-  confirmAccept: "비밀번호가 일치해요!",
-};
