--- src/core/signup/signupLabelText.ts
@@ -1,6 +1,0 @@
-export const SIGNUP_FIELD_LABEL = {
-  name: "이름",
-  email: "이메일",
-  password: "비밀번호",
-  confirm: "비밀번호 확인",
-};
